Rename folder from release to postbuild
