+++ packages/dnb-eufemia/package.json
@@ -22,12 +22,12 @@
     "build": "./scripts/prebuild/local-prebuild.sh",
     "prebuild:ci": "yarn build:prebuild",
     "build:ci": "yarn prebuild:ci && yarn postbuild:ci && yarn publish:prepare",
-    "postbuild:ci": "./scripts/release/postbuild.sh",
-    "build:cjs": "./scripts/release/babel-cjs.sh",
+    "postbuild:ci": "./scripts/postbuild/postbuild.sh",
+    "build:cjs": "./scripts/postbuild/babel-cjs.sh",
     "build:clean": "rm -rf build/**",
-    "build:copy": "./scripts/release/copy-build-artifacts.sh",
-    "build:es": "./scripts/release/babel-es.sh",
-    "build:esm": "./scripts/release/babel-esm.sh",
+    "build:copy": "./scripts/postbuild/copy-build-artifacts.sh",
+    "build:es": "./scripts/postbuild/babel-es.sh",
+    "build:esm": "./scripts/postbuild/babel-esm.sh",
     "prebuild:figma:ci": "yarn figma",
     "postbuild:figma:ci": "yarn test:update && yarn test:screenshots:update && babel-node ./scripts/figma/tasks/commitChanges",
     "build:pack": "yarn build && yarn publish:prepare && cd ./build && yarn pack",
@@ -37,7 +37,7 @@
     "build:types:definitions": "yarn tsc --project tsconfig.definitions.json",
     "build:types:dev": "nodemon --exec 'babel-node ./scripts/prebuild/generateTypes.js' --ext js --watch './src/**/*' --watch './scripts/**/*'",
     "build:types:esm": "yarn tsc --project tsconfig.definitions.json --declarationDir ./build/esm --rootDir src",
-    "build:umd": "./scripts/release/babel-umd.sh",
+    "build:umd": "./scripts/postbuild/babel-umd.sh",
     "precommit": "yarn lint-staged",
     "dev:icons": "nodemon --exec 'babel-node ./scripts/tools/convertIcons' --ignore '/icons/**' --ignore '*.json'",
     "dev:resources": "nodemon --exec 'babel-node ./scripts/prebuild/resources/makeResourcesPackage.js' --ext js,html,json,css,scss --watch './build/style/**/*' --watch './scripts/**/*' --ignore '*.json'",
@@ -59,9 +59,9 @@
     "lint:styles": "stylelint './src/**/*.{js,scss}'",
     "lint:styles:staged": "stylelint './{src,scripts}/**/*.{js,scss}'",
     "make-properties": "babel-node ./scripts/prebuild/makeProperties",
-    "publish:ci": "yarn publish:prepare && ./scripts/release/publish-release.sh",
+    "publish:ci": "yarn publish:prepare && ./scripts/postbuild/publish-release.sh",
     "publish:dry": "dotenv semantic-release --no-ci --dry-run",
-    "publish:prepare": "./scripts/release/publish-prepare.sh",
+    "publish:prepare": "./scripts/postbuild/publish-prepare.sh",
     "reset": "rm -rf ./node_modules ./components ./elements ./es ./icons ./extensions ./shared ./style ./umd",
     "skeleton:font": "nodemon --exec 'babel-node ./scripts/tools/createSkeletonFont.js'",
     "start": "start-storybook -s / -p 8002 --ci --no-release-notes --no-version-updates",
