+++ packages/dnb-eufemia/scripts/postbuild/__tests__/prepareForRelease.test.js
