+++ packages/dnb-eufemia/scripts/postbuild/babel-cjs.sh
@@ -19,8 +19,8 @@ echo 'Building cjs bundle done!'
 
 echo 'Copy additional files to cjs ...'
 
-OUT_DIR=./build/cjs babel-node ./scripts/release/copyTypeScriptFiles.js
-OUT_DIR=./build/cjs babel-node ./scripts/release/copyStyles.js
+OUT_DIR=./build/cjs babel-node ./scripts/postbuild/copyTypeScriptFiles.js
+OUT_DIR=./build/cjs babel-node ./scripts/postbuild/copyStyles.js
 
 echo 'Copy extra cjs package.json ...'
