+++ packages/dnb-eufemia/scripts/postbuild/babel-esm.sh
@@ -2,22 +2,22 @@
 
 set -e # Exit immediately if a command exits with a non-zero status.
 
-echo 'Building es bundle ...'
+echo 'Building esm bundle ...'
 
 cross-env \
 NODE_ENV=production \
-BABEL_ENV=es \
+BABEL_ENV=esm \
 babel ./src \
 --extensions '.js,.ts,.tsx' \
 --config-file ./babel.config.js \
---out-dir ./build/es \
+--out-dir ./build/esm \
 --copy-files \
 --no-copy-ignored \
 --ignore 'src/cjs,src/esm,src/umd,src/core,**/*.test.*,**/__tests__/**/*,**/*.stories.*,**/stories,**/*.d.ts'
 
-echo 'Building es bundle done!'
+echo 'Building esm bundle done!'
 
-echo 'Copy additional files to es ...'
+echo 'Copy additional files to esm ...'
 
-OUT_DIR=./build/es babel-node ./scripts/release/copyTypeScriptFiles.js
-OUT_DIR=./build/es babel-node ./scripts/release/copyStyles.js
+OUT_DIR=./build/esm babel-node ./scripts/postbuild/copyTypeScriptFiles.js
+OUT_DIR=./build/esm babel-node ./scripts/postbuild/copyStyles.js
