+++ packages/dnb-eufemia/scripts/postbuild/babel-umd.sh
