+++ packages/dnb-eufemia/scripts/postbuild/copyFinaleBuild.js
