+++ packages/dnb-eufemia/scripts/postbuild/prepareForRelease.js
