+++ packages/dnb-eufemia/scripts/postbuild/publish-prepare.sh
@@ -4,10 +4,10 @@ set -e # Exit immediately if a command exits with a non-zero status.
 
 echo 'Prepare before publish has started ...'
 
-babel-node ./scripts/release/prepareForRelease.js
+babel-node ./scripts/postbuild/prepareForRelease.js
 
 echo 'Testing the postbuild before publish ...'
 
-jest --ci --rootDir ./scripts/release ./prepareForRelease.test.js
+jest --ci --rootDir ./scripts/postbuild ./prepareForRelease.test.js
 
 echo 'Prepare before publish is done!'
