+++ packages/dnb-eufemia/scripts/postbuild/publish-release.sh
